需求#[23769]：【摄像机任务】缺陷测点，选择告警等级，下发任务生成缺陷数据，所有缺陷数据的等级都为一般告警
    [描述说明]:完善缺陷信息处理逻辑
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-accessvideo/src/main/java/com/yjh/accessvideo/module/device/dao/AnalyseDataOperateDao.java
@@ -42,6 +42,6 @@ public interface AnalyseDataOperateDao {
     Map<String,Object> selectWarnInfo(@Param(value = "deviceMeteId")Long deviceMeteId);
 
     //根据缺陷类型查询缺陷等级
-    int selectAlgorithmDefectInfo(@Param(value = "defectType")Long defectType);
+    int selectAlgorithmDefectInfo(@Param(value = "defectType")Integer defectType);
 
 }
